requirements-dev.txt
# GLOBAL REQUIREMENTS
-r requirements.txt

# SOFT DEPENDENCIES
matplotlib

# SETUP RELATED
setuptools_scm

# FOR DOCUMENTATION
sphinx>=7.3
numpydoc
sphinx_design
sphinx_numfig
sphinx_gallery>=0.16
pydata_sphinx_theme
sphinx_automodapi
ipykernel
pooch

# FOR TESTING
flake8
pytest
coveralls
pytest_cov

#RUN DARTS
open-darts
open-darts-flash 